修改内容

+++ src/api_database/api_auto_shutter/auto_shutter_api.rs
@@ -3,55 +3,6 @@ use napi_derive::napi;
 
 use crate::error_handle::err_handle::*;
 use crate::service_database::database_auto_shutter::service::auto_shutter_service::*;
-// use crate::service_database::database_cache::service::auto_shutter_cache_service as cache_service;
-
-// #[napi(namespace = "autoShutter")]
-// pub async fn sync_cache_to_local_api() -> Result<()> {
-//     let sync_data = cache_service::get_all_model_auto_shutter_entity_cache()
-//         .await
-//         .map_err(SyncError::CacheReadError)?;
-
-//     if sync_data.is_empty() {
-//         eprintln!("⚠️ [同步] 缓存数据库无数据可同步到本地");
-//         return Ok(());
-//     }
-
-//     // 假设 local_service 接受引用或 clone，保持你原逻辑
-//     local_service::read_current_model_auto_shutter_entity(sync_data.clone())
-//         .await
-//         .map_err(SyncError::LocalWriteError)?;
-
-//     eprintln!("✅ [同步] 成功同步 {} 条数据到本地数据库", sync_data.len());
-//     Ok(())
-// }
-
-// #[napi(namespace = "autoShutter")]
-// pub async fn sync_local_to_cache_api() -> Result<()> {
-//     let sync_data = local_service::get_current_all_model_auto_shutter_entity()
-//         .await
-//         .map_err(SyncError::LocalReadError)?;
-
-//     if sync_data.is_empty() {
-//         eprintln!("⚠️ [回写] 本地数据库无数据可回写缓存");
-//         return Ok(());
-//     }
-
-//     cache_service::read_model_auto_shutter_entity_cache(sync_data.clone())
-//         .await
-//         .map_err(SyncError::CacheWriteError)?;
-
-//     eprintln!("✅ [回写] 成功回写 {} 条数据到缓存数据库", sync_data.len());
-//     Ok(())
-// }
-
-// #[napi(namespace = "autoShutter")]
-
-// pub async fn read_auto_shutter_cache_api(data: Vec<FullCacheData>) -> Result<()> {
-//     read_model_auto_shutter_entity_cache(data)
-//         .await
-//         .map_err(handle_db_err)?;
-//     Ok(())
-// }
 
 #[napi(namespace = "autoShutter")]
 /// 更新自动快照 (Service: update_model_auto_shutter_entity_cache)
@@ -62,15 +13,6 @@ pub async fn update_auto_shutter_cache_api(data: AutoShutterData, model_id: Stri
     Ok(result)
 }
 
-// #[napi(namespace = "autoShutter")]
-// /// 查询缓存数据库中模型的数据
-// pub async fn get_all_shutter_cache_api_model_id(model_id: String) -> Result<Vec<FullCacheData>> {
-//     let result = get_all_model_auto_shutter_entity_cache_model_id(model_id)
-//         .await
-//         .map_err(handle_db_err)?;
-//     Ok(result)
-// }
-
 #[napi(namespace = "autoShutter")]
 /// 获取快照列表 (Service: get_all_model_auto_shutter_entity_list_cache)
 pub async fn get_auto_shutter_cache_list_api(
@@ -79,10 +21,9 @@ pub async fn get_auto_shutter_cache_list_api(
     model_id: String,
 ) -> Result<Vec<AutoShutterListItem>> {
     // 注意：服务层需要 u64，这里进行转换
-    let result =
-        get_all_model_auto_shutter_entity_list_cache(order_flag, auto_count as u64, model_id)
-            .await
-            .map_err(handle_db_err)?;
+    let result = get_all_model_auto_shutter_entity_list_cache(order_flag, auto_count, model_id)
+        .await
+        .map_err(handle_db_err)?;
 
     Ok(result)
 }
